devices/requirements.txt
pip
paho-mqtt